style: Home 페이지에 나타나기 애니메이션 적용
- About 컴포넌트 각 섹션 파일로 분리
